requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.1.6
django-ninja==1.3.0
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
sqlparse==0.5.3
typing_extensions==4.12.2
